Fix SonarApiClient Bug
Sonar saves some projects' component keys as "owner:projectName".
We used to always use the projectName as component key so even though the analysis run correctly, when we were calling the sonar api we were getting empty responses.

What I did was that I make a check to find the correct componentKey

I found that after this both the "component not found" and the code smell, td etc populations started to work correctly