Refactor Analysis Service (1)
First of all we moved most of the login inside the AnalysisService so the controller cares only about getting the request and delegating the work to the AnalysisService.

Also, we implemented an Exception Handling mechanism to reduce the ifs. The AnalysisExeption handler uses a decision table to determine what kind of response to return based on the exception